+++ CHANGELOG.md
@@ -49,6 +49,9 @@ MINOR
 ### PaymentSheet
 * [Added] Added `financialConnectionsPermissions` to `LinkConfiguration`, allowing `LinkControllerPreview` users to request Financial Connections data permissions (private preview).
 
+### Financial Connections
+* [Fixed] Preserved `no_eligible_accounts` in `onEvent` error callbacks instead of reporting it as `unexpected_error`.
+
 ## 26.7.0 2026-08-17
 ### PaymentSheet
 * [Fixed] Fixed FlowController not presenting the Link sheet when tapping Link in the horizontal payment method layout.